aggiunte repository e endpoint per WaterLimit

+++ codice/OnlyFarms/OnlyFarms.Core/Models/Crop.cs
@@ -4,13 +4,13 @@ public record Crop : IHasId     // entita' Coltivazione
 {
     public int Id { get; set; }
     public required string Tag { get; set; }
-    public required float SurfaceArea { get; set; }
+    public required int SurfaceArea { get; set; }
     
     [JsonConverter(typeof(JsonStringEnumConverter))]    // per poter avere il parametro rappresentato come stringa e non come int nei json della REST API (nel DB viene comunque salvato come int)
     public required IrrigationType IrrigationType { get; set; }
     
-    public required float WaterNeeds { get; set; }
-    public required float IdealHumidity { get; set; }
+    public required int WaterNeeds { get; set; }
+    public required int IdealHumidity { get; set; }
     
     public int FarmingCompanyId { get; set; }   // riferimento all'azienda a cui appartiene
